refactored dependency_analysis_types into new header `common`.
